Migracion de proyectos a linux > adición de archivos script para compilacion

+++ pic12f683/02_pwm/02_pwm.c
@@ -18,7 +18,7 @@
 const unsigned short MAX_PWM_VALUE = 280;
 const unsigned short DELTA_PWM = 28;
 
-int pwm_value;
+unsigned short pwm_value;
 bool increment;
 
 void set_duty_cycle( unsigned short value) {
